+++ src/api/files.js
@@ -40,13 +40,13 @@ export async function getFoldersByFolder(parent) {
 export async function fetchSharedFiles() {
   const res = await Axios.get(`${baseURL}/api/files?shares&appId=drive`);
   const files = res.data;
-  return files;
+  return files.successful;
 }
 
 export async function fetchExternalTransferdFiles() {
   const res = await Axios.get(`${baseURL}/api/files?shares&appId=dropbox`);
   const files = res.data;
-  return files;
+  return files.successful;
 }
 
 /**