+++ CHANGELOG.md
@@ -5,8 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v3.1.1]
 
+## Added
+
+- info about the approvers in the ui.
+## [v2.0.2] - 2020-10-28
+
+### Removed
+
+- default timoout for the axios request because some big files uploads that took more then 10 sec faild.
 ## [v2.3.0] - 2020-12-24
 ### Added